Create README.md
Here, there is a code to simulate a light curve due to a double white dwarf (DWD) system which is DWDcase.cpp
Also to evaulate the amplitudes of ellipsoidal variation and Doppler boosting in these systems, 
I develope two python-based codes:  Ellipsoidal.py and Doppler.py
To study a time-series data points taken by the TESS telescope a python code in Jupyter environment was developed.  
Two files include (i) the TESS throughtput function for the T-band filter versus wavelength, and (ii)
the limb-darkening and gravity-darkening coefficients for WDs in the TESS T-band for various values of their temperature and surface gravity.